Serial Communication, with Arduino
The Arduino receives bytes, usually 6 bytes, which comprise one command. Each byte string is 6 bytes including ';' terminating byte. First byte is command byte, 2-3rd byte is parameter 1 an unsigned long which is 8 bytes on an Arduino, 4-5 byte is parameter 2 an unsigned long which is 8 bytes on an Arduino.

Also, there is a command buffer, which can currently hold 8 command byte strings, but we can easily extend that to a hundred.